// src/api/company/company.interface.ts
export interface ICompany {
  id: string;
  short_id: string;
  name: string;
  code: string;
  fullName: string;
  site: string;
  readonly app_key: string;
  updated_at: Date;
  created_at: Date;
}


  // "site":"http://буревестник.рф",
  // "code": "BRVST",
  // "app_key": "90c43df89760a5095ef6",
  // "updated_at": "2018-10-18T20:10:28.698Z",
  // "name": "Буревестник",
  // "created_at": "2018-10-18T20:10:28.698Z",
  // "fullName": "ООО Буревестник",
  // "id": "db9993a0-d311-11e8-8403-8730a7b86236",
  // "short_id": "5UP9QB"